frontend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
requests
python-dotenv
openai
google-generativeai
